requirements.txt
torch
torchvision
numpy
Pillow
opencv-python
scikit-learn
pandas
PyYAML
